refactor: flatten _layout directories to _layout.tsx files
Move layout route files from _layout/_layout.tsx to flat _layout.tsx
for tasks, settings, and users. Colocated files (+components, +hooks,
+queries) moved to parent directory with import paths updated.
