+++ app/routes/_authenticated/users/+components/data-table-toolbar.tsx
@@ -2,7 +2,7 @@ import { Cross2Icon } from '@radix-ui/react-icons'
 import type { Table } from '@tanstack/react-table'
 import { Button } from '~/components/ui/button'
 import { useDataTableState } from '../+hooks/use-data-table-state'
-import { userTypes } from '../../+shared/data/data'
+import { userTypes } from '../+shared/data/data'
 import { DataTableFacetedFilter } from './data-table-faceted-filter'
 import { DataTableViewOptions } from './data-table-view-options'
 import { SearchInput } from './search-input'
